requirements.txt
anyio==4.13.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2026.2.25
charset-normalizer==3.4.7
flake8==7.1.2
flake8-isort==6.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.11
isort==6.0.1
mccabe==0.7.0
pycodestyle==2.12.1
pyflakes==3.2.0
pyTelegramBotAPI==4.14.1
python-dotenv==1.2.2
python-telegram-bot==20.7
pytz==2026.1.post1
requests==2.33.1
setuptools==82.0.1
six==1.17.0
sniffio==1.3.1
tornado==6.5.5
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
